\name{circleplot_fun}
\alias{circleplot}
\title{Plot a distance matrix}
\usage{
circleplot(distance.matrix, plot.control)
}
\arguments{
  \item{distance.matrix}{a distance matrix (class dist) containing either continuous or binary data}

  \item{plot.control}{a list giving information on how the plot should appear. If given, may contain any or all of the named objects given in notes (below).}
}
\value{
Draws a plot showing the labels of the distance matrix as nodes around the circumference of a circle, with quadratic lines joining connected nodes.
}
\description{
Draws a circular plot of a distance matrix. 
}
\note{
options for plot.control are as follows:
  \describe{
    \item{points}{a data.frame giving point attributes, with colnames=c('label', "colour", "size"). Can be generated by point.attr().}
    \item{line.gradient}{logical, default to FALSE; should lines display a colour gradient. This is usually only sensible for binary matrices (i.e. where line width or colour is not meaningful).}
    \item{line.breaks} vector of breaks in distance.matrix used to decide line colours}
    \item{line.colours} vector containing colours, either named or in hexadecimal. Used with 'line.breaks' above; length should equal length(breaks)-1}
    \item{line.width}{vector. if length-1 contains absolute value; if length-2 contains minimum and max}
    \item{line.curvature}{the percentage of maximum (as given by line.width above) sent to gaussian curve i.e. curvature=1 always has a min line width =0, max line width=max, while curvature=0 is a line with constant width}
}
}

